refactor(inventory): split controller into per-module routers (orchestration prep)
Each RFC-0061 module now owns src/controllers/inventory/<module>.controller.ts
(M1..M9 + meta), composed by inventory.controller.ts; cross-cutting guards
(uuid, Idempotency-Key, confirmation token, 501 defer) move to
inventory/shared.ts. Behavior is unchanged (typecheck + 40 contract/DTO tests
green). This lets module implementations land as parallel PRs that never touch
each other's files.
